homework/tkinter_example/requirements.txt
matplotlib==3.5.2
opencv_python==4.5.5.64
Pillow==9.1.1
pygame==2.1.2
